Agregar codigo javascript para mostrar las tareas.

+++ src/main/java/com/Hursarior/Alura/ToDoApp/Service/TaskService.java
@@ -9,33 +9,54 @@ import com.Hursarior.Alura.ToDoApp.Persistence.Entity.task_status;
 import com.Hursarior.Alura.ToDoApp.Persistence.Entity.Repository.TaskRepository;
 import com.Hursarior.Alura.ToDoApp.Service.DTO.TaskDTO;
 
+import jakarta.persistence.EntityNotFoundException;
+import jakarta.transaction.Transactional;
+
 @Service
 public class TaskService {
-    
+
     private final TaskRepository repository;
 
     public TaskService(TaskRepository repository) {
         this.repository = repository;
     }
 
-    public Task CrearTask(TaskDTO taskDTO){
-         return this.repository.save(new Task(taskDTO));
-        
+    public Task CrearTask(TaskDTO taskDTO) {
+        return this.repository.save(new Task(taskDTO));
+
     }
 
-    public List<Task> ObtenerTask(){
+    public List<Task> ObtenerTask() {
         return this.repository.findAll();
-        
-    }
-
-    public void DeleteTask(Long id){
-
-       Task task = this.repository.getReferenceById(id);
-       this.repository.delete(task);
-    }
-
-    public List<Task> filtrarStatus(task_status status){
-        return this.repository.findallByStatus(status);
 
     }
+
+    public void DeleteTask(Long id) {
+        if (this.repository.existsById(id)) {
+            Task task = this.repository.getReferenceById(id);
+            this.repository.delete(task);
+        } else {
+            throw new EntityNotFoundException();
+        }
+
+    }
+
+    public List<Task> filtrarStatus(task_status status) {
+        if (this.repository.existsByStatus(status)) {
+            return this.repository.findallByStatus(status);
+        }
+        throw new EntityNotFoundException();
+    }
+
+    @Transactional
+    public boolean UpdateTaskFinished(Long id) {
+        if (this.repository.existsById(id)) {
+            this.repository.markTaskFinish(id);
+            return true;
+        } else {
+            return false;
+        }
+
+    }
+
 }
